TestApp/requirements.txt
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.4.0
pandas==2.2.3
matplotlib==3.7.2
psutil==5.9.5
python-dotenv==1.0.0
cryptography>=40.0.0
